+++ srai/joiners/base.py
@@ -9,7 +9,7 @@ class BaseJoiner(abc.ABC):
     """Base abstract class for joiners."""
 
     @abc.abstractmethod
-    def join(self, regions: gpd.GeoDataFrame, features: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
+    def transform(self, regions: gpd.GeoDataFrame, features: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
         """
         Join features to regions.